Added 'About Us' section content and assets
In this commit, I added content and assets to the "About Us" section in the HTML file. The section now includes a descriptive subtitle in both English and Japanese, a mission statement, and a brief description of our commitment to providing the Japanese mindset of hospitality, known as Omotenashi, to our customers. I also included an image of sushi (sourced from 'sushi-2.png') within the section.

These changes aim to provide users with insights into our mission and values, enhancing the overall content and user experience of the "About Us" section on the website.